[refactor] :recycle: refactor bff folders structure
