+++ src/app/api/hooks/use-nft.ts
@@ -1,4 +1,4 @@
-import { endpoints } from 'src/app/api/_hooks/api-resolver'
+import { endpoints } from 'src/app/api/hooks/api-resolver'
 
 import { useGetCommon } from './common'
