+++ src/app/api/middleware/validators/base-security-validator.ts
@@ -1,10 +1,12 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { getIpFromRequest } from 'src/app/api/_utils/request-utils'
-import { extractjwtTokenFromHeader } from 'src/app/api/_utils/jwt-utils'
-import { checkUserHaveActiveSession } from 'src/app/api/_data/data-service'
+import { getIpFromRequest } from 'src/app/api/middleware/utils/network-utils'
+import { extractjwtTokenFromHeader } from 'src/app/api/middleware/utils/jwt-utils'
+import { checkUserHaveActiveSession } from 'src/app/api/services/db/chatterpay-db-service'
 
 import { JwtPayload } from 'src/types/jwt'
+
+// ----------------------------------------------------------------------
 
 type AuthContext = {
   userId: string
@@ -12,11 +14,14 @@ type AuthContext = {
   jwtToken: JwtPayload
 }
 
+// ----------------------------------------------------------------------
+
 export async function validateRequestSecurity(
   req: NextRequest,
   userId: string
 ): Promise<AuthContext | NextResponse> {
   const ip = getIpFromRequest(req)
+
   const jwtTokenDecoded: JwtPayload | null = extractjwtTokenFromHeader(
     req.headers.get('Authorization')
   )
