+++ src/app/api/services/google/recaptcha-service.ts
@@ -1,27 +1,13 @@
-import { NextRequest } from 'next/server'
-
 import { UI_BASE_URL, RECAPTCHA_API_KEY } from 'src/config-global'
 
 import { RecaptchaResponse } from 'src/types/recaptcha'
 
 // ----------------------------------------------------------------------
-export function getIpFromRequest(req: NextRequest) {
-  const forwarded = req.headers.get('x-forwarded-for')
-  let ip = '0.0.0.0'
-  if (typeof forwarded === 'string') {
-    ip = forwarded.split(',')[0]
-  } else {
-    ip = req.ip || '0.0.0.0'
-  }
-
-  return ip
-}
 
 export async function validateRecaptcha(
   recaptchaToken: string,
   ip: string = '0.0.0.0'
 ): Promise<RecaptchaResponse> {
-  // Validar el token de recaptcha
   const verifyEndpoint = 'https://www.google.com/recaptcha/api/siteverify'
   const recaptchaResponse = await fetch(verifyEndpoint, {
     method: 'POST',
