+++ src/app/api/v1/auth/code/route.ts
@@ -1,8 +1,14 @@
 import { createHash } from 'crypto'
 import { NextRequest, NextResponse } from 'next/server'
 
-import { getIpFromRequest, validateRecaptcha } from 'src/app/api/_utils/request-utils'
-import { getUserByPhone, updateUserCode, createUserSession } from 'src/app/api/_data/data-service'
+import { getIpFromRequest } from 'src/app/api/middleware/utils/network-utils'
+import { send2FACode } from 'src/app/api/services/chatizalo/chatizalo-service'
+import { validateRecaptcha } from 'src/app/api/services/google/recaptcha-service'
+import {
+  getUserByPhone,
+  updateUserCode,
+  createUserSession
+} from 'src/app/api/services/db/chatterpay-db-service'
 import {
   BOT_API_URL,
   BOT_API_TOKEN,
@@ -13,7 +19,6 @@ import {
 
 import { IAccount } from 'src/types/account'
 
-import { send2FACode } from '../../_common/chatizalo'
 // ----------------------------------------------------------------------
 
 export async function POST(req: NextRequest) {
