+++ src/app/api/v1/auth/login/route.ts
@@ -1,13 +1,14 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { generateJwtToken } from 'src/app/api/_utils/jwt-utils'
-import { getIpFromRequest, validateRecaptcha } from 'src/app/api/_utils/request-utils'
+import { generateJwtToken } from 'src/app/api/middleware/utils/jwt-utils'
+import { getIpFromRequest } from 'src/app/api/middleware/utils/network-utils'
+import { validateRecaptcha } from 'src/app/api/services/google/recaptcha-service'
 import {
   getUserByPhone,
   updateUserCode,
   updateUserSessionStatus,
   validateUserHave1SessionCreated
-} from 'src/app/api/_data/data-service'
+} from 'src/app/api/services/db/chatterpay-db-service'
 
 import { jwtPayloadUser } from 'src/types/jwt'
 import { IAccount, UserSession } from 'src/types/account'
