+++ src/app/api/v1/health/route.ts
@@ -1,5 +1,7 @@
 import { NextResponse } from 'next/server'
 
+// ----------------------------------------------------------------------
+
 export async function GET() {
   return NextResponse.json({ data: 'ok' })
 }
