+++ src/app/api/v1/nft/[id]/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server'
 
-import { getNftById } from 'src/app/api/_data/data-service'
+import { getNftById } from 'src/app/api/services/db/chatterpay-db-service'
 
 import { INFT } from 'src/types/wallet'
 import { IErrorResponse } from 'src/types/api'
