+++ src/app/api/v1/user/[id]/email/route.ts
@@ -1,11 +1,14 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { getUserById, updateUserCode, updateUserEmail } from 'src/app/api/_data/data-service'
+import { validateRequestSecurity } from 'src/app/api/middleware/validators/base-security-validator'
+import { validateUserCommonsInputs } from 'src/app/api/middleware/validators/user-common-inputs-validator'
+import {
+  getUserById,
+  updateUserCode,
+  updateUserEmail
+} from 'src/app/api/services/db/chatterpay-db-service'
 
 import { IAccount } from 'src/types/account'
-
-import { validateUserCommonsInputs } from '../../userCommonInputsValidator'
-import { validateRequestSecurity } from '../../../_common/baseSecurityRoute'
 
 // ----------------------------------------------------------------------
