+++ src/app/api/v1/user/[id]/transfer-all/route.ts
@@ -1,12 +1,11 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { transferAll } from 'src/app/api/_data/blk-service'
-import { getUserById } from 'src/app/api/_data/data-service'
+import { getUserById } from 'src/app/api/services/db/chatterpay-db-service'
+import { transferAll } from 'src/app/api/services/blockchain/blockchain-service'
+import { validateRequestSecurity } from 'src/app/api/middleware/validators/base-security-validator'
+import { validateUserCommonsInputs } from 'src/app/api/middleware/validators/user-common-inputs-validator'
 
 import { IAccount } from 'src/types/account'
-
-import { validateUserCommonsInputs } from '../../userCommonInputsValidator'
-import { validateRequestSecurity } from '../../../_common/baseSecurityRoute'
 
 // ----------------------------------------------------------------------
