+++ src/app/api/v1/wallet/[id]/nfts/[nft_id]/route.ts
@@ -1,11 +1,10 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { getWalletNft } from 'src/app/api/_data/data-service'
+import { getWalletNft } from 'src/app/api/services/db/chatterpay-db-service'
+import { validateRequestSecurity } from 'src/app/api/middleware/validators/base-security-validator'
+import { validateWalletCommonsInputs as validateWalletCommonInputs } from 'src/app/api/middleware/validators/wallet-common-inputs-validator'
 
 import { IErrorResponse } from 'src/types/api'
-
-import { validateRequestSecurity } from '../../../../_common/baseSecurityRoute'
-import { validateWalletCommonsInputs as validateWalletCommonInputs } from '../../../walletCommonInputsValidator'
 
 // ----------------------------------------------------------------------
